checkpoint: renames "Config/types/Config_ParsingError.ts" to "Config/ParsingError.ts"
