+++ src/features/TextToImage/Config/types/exports.ts
@@ -1,3 +1,3 @@
 export {
   TextToImage_Config_ParsingError as Config_ParsingError,
-} from './Config_ParsingError';
+} from '../ParsingError';
